Update docker image to php7.3
Updated Composer to v1.9.1.
Due to the container now using Debian10 (instead of 9)
we had to set the package manager preference when it installs
nodejs since the version that comes with Debian10 is node10
and it seems it has some issues with npm not being installed.
Closes #290
Increased memory limit of php scripts due to feature tests.
Closes #289